move inbox related files to TS files folder
